+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 NOTE: all releases may include dependency updates, not specifically mentioned
 
-UNRELEASED
+## UNRELEASED
 
+- feat: integrate try-as library [#912](https://github.com/hypermodeinc/modus/pull/912)
 - fix: check topic actor status before publishing events [#918](https://github.com/hypermodeinc/modus/pull/918)
 
 ## 2025-06-25 - Runtime v0.18.2